changed package name to match sample name
