+++ src/main/java/com/esri/samples/na/service_area_task/ServiceAreaTask.png
