+++ src/main/resources/fxml/service_area_task.fxml
@@ -23,7 +23,7 @@
 <?import javafx.scene.layout.HBox?>
 <?import com.esri.arcgisruntime.mapping.view.MapView?>
 
-<StackPane  fx:controller="com.esri.samples.na.service_areas.ServiceAreaTaskController" xmlns:fx="http://javafx.com/fxml"
+<StackPane fx:controller="com.esri.samples.na.service_area_task.ServiceAreaTaskController" xmlns:fx="http://javafx.com/fxml"
            stylesheets="/css/style.css">
            
     <!--SDK SceneView-->
